… 1.5/Assemblies/Celsius.dll
… Source/TemperatureInfo.cs