Hallo!

I like to build base and have big map. I noticed a while back that Celsius didn't seem to scale all that well with larger map sizes and got curious how it did its thing.
From that this change was created, which I've been using successfully in both "standalone" and a larger (~200 mods) modpack.

## Changes
- Explicit Array.Clear rather than creating a new array for thermalProps. Afaik this works out to the same code once the compiler ran over it, but I personally dislike these rather hidden and unexpected optimizations. I also don't know how Mono plays into it
- The `#if DEBUG` conditional compilation was removed to be able to actually activate debug logging at runtime without recompiling
- TicksPerSlice, SliceCount, slice and so on were removed and instead the whole map is updated all at once
- The whole-map update uses a worksetTemperature array to prevent a single temperature change to propagate across the whole map in a single tick
- The whole-map update takes place once every TicksPerUpdate, rather than 4 times and every TicksPerSlice (which was 1/4th of TicksPerUpdate)
- TicksPerUpdate can now be upped to 1000 ticks, since I like to play on speed 4, which aims at 900 TPS. Default wasn't adjusted
- The rare update stuff (like colour and thermal properties) were moved to be changed on every iteration, which works out to the same timing as previously, just without the slice stuff
- The temperature update calculation was moved into a background thread which is synchronized to the ticking on the main thread
- The temperature update was parallelized (pretty naively with a `Parallel.ForEach`)
- Some comments were added (half of which so I could make sense of what was happening when)

## Results
- Before (on a 600x600 cell map): Average per-tick timing was 1,2ms, however the tick the computation happened on took ~10ms, and the rare update took 150ms -> Some stuttering when the update happens and higher speeds or larger maps are pretty bad. Max speed was 400 TPS
- After: Average per-tick timing is 0ms, and the background thread computing the whole map takes ~8ms (albeit on my 5950X, so less powerful CPUs will likely take longer) -> *No* impact on the game simulation, max speed is 900 TPS (maximum)

## Questions
I'm new to Rimworld modding and know Rimworld and its creators don't really like Multithreading. I tested these changes extensively though, but:
- I haven't caught any errors regarding synchronization. From my understanding all the code does is read from Rimworld structures, then (asynchronously) update its own temperature array, then provide that array to Rimworld (on the main thread again). This should be fine and should not cause race conditions, but in particular the snow melting/freezing seems to work but I'm not sure about.
- I've tried continuing to play on an older save with a large modpack (~200 mods) and the game crashed after loading the save. I couldn't find any errors except the game process literally froze. I tested some more and while I couldn't find any reason for it to be happening, and the save was kinda skuffed already, I do want to mention it
- The default update interval (120 ticks) is pretty hard on the CPU when you play at speed 4 (since it's essentially 900 / 120 = ~7.3 updates a second). That's a lot, even parallelized. I would propose upping it, or maybe changing it to be an updates per second/seconds per update setting rather than per ticks. We could still use ticks internally
- I tried to test a lot but obviously couldn't test everything. I'm not sure what happens when you raid another settlement or what happens on Anomaly-specific maps (I don't have that DLC)

## Further work
- Periodically clearing the thermal properties and room temperatures isn't optimal. It would be better to either only clear part of it, and not delete the dictionary completely and instead reset the values to a default value (like `float.MinValue`)
- I think some of the stuff from Rimworld can be cached and used to optimize the data structures, like cells and cell indices. It's not like we're suffering for memory usually, but I'm not sure how they are handled internally by Rimworld, and how it could affect some things (like embarking) or how much performance it would improve

I'm happy to discuss this change with you and I'd be fine if you would want to make it a setting or fork or something like that. I haven't published it to the Workshop, just forked for the code commit.