@L3tum 2024-08-13
I just noticed that I think from the change to the worksetTemperature there's now an issue with essentially "infinite energy" being pulled, since the calculation takes `temperatures` rather than `worksetTemperatures` into account. I need to think a second on how it could be mitigated without impacting the multithreading too much, so I converted it into a draft. My first instinct would be to, instead of having a worksetTemperature, save the temperatureChange and then cumulatively applying it. It would reduce performance a bit but not by much. I'm not all that well versed in energy simulation. If you have any input or any sources I could read I'd appreciate it. Most of what I can find is highly scientific and less about the nitty gritty of performantly simulating it.

The few things I've found, i.e. [1](https://github.com/gkigiermo/parallel-heat-transfer/tree/main) and [2](https://github.com/scottsumrall/Parallel-Heat-Sim-MPI) seem to simply accept the error as is from what I could tell. That feels a bit weird although is probably acceptable for a game.