@L3tum 2024-08-13
lol, disregard the above. I thought of it as a discrete energy transfer rather than simultanously solving the same function on multiple inputs. I got spooked by the temperature on a testmap steadily rising, but it reached an equilibirum after ~2 ingame hours. I'm gonna add another comment to the code to explain this as well just to make sure. At least this also means we could get another significant speed-up with some SIMD acceleration, especially around the terrain code I guess.

That mentioned, I did also try to have the different components (terrain, cell, ambient) in different functions so that they could be throttled individually as well, but I found it made the code a whole lot more unreadable. At least the melting code could be separated, but I'll let you look over this first to see what you think of it.