@GarwelGarwel 2024-08-16
Hey, this is a great effort, and something I wanted but never had the heart to do. I have almost no experience with multi-thread programming. All I know about it is that it may be a hell to debug. But if the code works well, it's a huge improvement.

Don't forget to mark the PR as ready when it is, so that I could test it and see how best to integrate it.