@L3tum 2024-08-19
Hey @GarwelGarwel , thank you!
I had a few issues with a borked Rimworld installation, but after reinstalling I could test the ice/snow melting and it works for both terrain and cell and replaces an ice sheet with water. That was the major thing that I didn't know if it would work accurately in multithreading, but it does, so the PR for me is gtg and I'll undraft it in a sec.
Regarding SIMD I noticed Unity is still using Mono, which doesn't support the newer SIMD stuff that .NET Core 3 and later put out. Honestly thought it had already moved over at least for desktop, since they published plans for that a few years ago. So that's unfortunately a no-go for now, afaik